requirements.txt
Flask==2.2.3
pymongo==4.3.3
python-dotenv==1.0.0
pyarrow>=19.0.1
flask-caching==2.0.2
flask-cors==4.0.1
Werkzeug==2.2.2
pyzmq==25.1.2
pandas>=1.5.0

